Feature/auth flow (#131)
Auth Context which stores auth status and user object
