+++ source/App.js
@@ -8,10 +8,10 @@
 
 import React, { Component } from 'react';
 import Config from 'react-native-config';
-import Navigator from './navigator';
 import StorybookUIRoot from '../storybook/index';
-
+import { AuthProvider } from './store/AuthContext';
 import StoreContext from './helpers/StoreContext';
+import Navigator from './navigator';
 import StorageService, { COMPLETED_FORMS_KEY } from './services/StorageService';
 
 /**
@@ -45,16 +45,20 @@ export default class App extends Component {
   render() {
     if (Config.IS_STORYBOOK === 'true') {
       return (
-        <StoreContext.Provider value={this.state}>
-          <StorybookUIRoot />
-        </StoreContext.Provider>
+        <AuthProvider>
+          <StoreContext.Provider value={this.state}>
+            <StorybookUIRoot />
+          </StoreContext.Provider>
+        </AuthProvider>
       );
     }
 
     return (
-      <StoreContext.Provider value={this.state}>
-        <Navigator />
-      </StoreContext.Provider>
+      <AuthProvider>
+        <StoreContext.Provider value={this.state}>
+          <Navigator />
+        </StoreContext.Provider>
+      </AuthProvider>
     );
   }
 }
