+++ source/components/organisms/index.js
@@ -1,5 +1,4 @@
 export { default as Chat } from './Chat';
 export { default as FormAgent } from './FormAgent';
 export { default as WatsonAgent } from './WatsonAgent';
-export { default as withAuthentication } from './withAuthentication';
 export { default as withForm } from './withForm';
