--- source/components/organisms/withAuthentication/index.js
@@ -1,1 +1,0 @@
-export { default } from './withAuthentication';
